Modifying class and function for DistributedDataParallel.
